Move sections 7 & 8 one back
This opens up a new section before "Error recovery" to teach about
fallible actions (defined with `=>?`).
