+++ doc/calculator/src/calculator7.lalrpop
@@ -1,21 +1,33 @@
 use std::str::FromStr;
 use ast::{Expr, Opcode};
+use lalrpop_util::ErrorRecovery;
 
-grammar(scale: i32);
+grammar<'err>(errors: &'err mut Vec<ErrorRecovery<usize, Token<'input>, &'static str>>);
 
-pub Expr: Box<Expr> = { // (1)
-    Expr ExprOp Factor => Box::new(Expr::Op(<>)), // (2)
-    Factor,
+pub Exprs = Comma<Expr>;
+
+Comma<T>: Vec<T> = { // (0)
+    <v:(<T> ",")*> <e:T?> => match e { // (1)
+        None=> v,
+        Some(e) => {
+            let mut v = v;
+            v.push(e);
+            v
+        }
+    }
 };
+
+Tier<Op,NextTier>: Box<Expr> = {
+    Tier<Op,NextTier> Op NextTier => Box::new(Expr::Op(<>)),
+    NextTier
+};
+
+Expr = Tier<ExprOp, Factor>;
+Factor = Tier<FactorOp, Term>;
 
 ExprOp: Opcode = { // (3)
     "+" => Opcode::Add,
     "-" => Opcode::Sub,
-};
-
-Factor: Box<Expr> = {
-    Factor FactorOp Term => Box::new(Expr::Op(<>)),
-    Term,
 };
 
 FactorOp: Opcode = {
@@ -25,10 +37,11 @@ FactorOp: Opcode = {
 
 Term: Box<Expr> = {
     Num => Box::new(Expr::Number(<>)),
-    "(" <Expr> ")"
+    "(" <Expr> ")",
+    ! => { errors.push(<>); Box::new(Expr::Error) },
 };
 
 Num: i32 = {
-    r"[0-9]+" => i32::from_str(<>).unwrap()*scale,
+    r"[0-9]+" => i32::from_str(<>).unwrap()
 };
