+++ doc/calculator/src/calculator8.lalrpop
@@ -1,36 +1,34 @@
-use ast::{ExprSymbol, Opcode};
-use tok8::Tok;
+use std::str::FromStr;
+use ast::{Expr, Opcode};
 
-grammar<'input>(input: &'input str);
+grammar(scale: i32);
 
-pub Expr: Box<ExprSymbol<'input>> = { // (1)
-    Expr "ExprOp" Factor => Box::new(ExprSymbol::Op(<>)), // (2)
+pub Expr: Box<Expr> = { // (1)
+    Expr ExprOp Factor => Box::new(Expr::Op(<>)), // (2)
     Factor,
 };
 
-Factor: Box<ExprSymbol<'input>> = {
-    Factor "FactorOp" Term => Box::new(ExprSymbol::Op(<>)),
+ExprOp: Opcode = { // (3)
+    "+" => Opcode::Add,
+    "-" => Opcode::Sub,
+};
+
+Factor: Box<Expr> = {
+    Factor FactorOp Term => Box::new(Expr::Op(<>)),
     Term,
 };
 
+FactorOp: Opcode = {
+    "*" => Opcode::Mul,
+    "/" => Opcode::Div,
+};
 
-Term: Box<ExprSymbol<'input>> = {
-    "num" => Box::new(ExprSymbol::NumSymbol(<>)),
+Term: Box<Expr> = {
+    Num => Box::new(Expr::Number(<>)),
     "(" <Expr> ")"
 };
 
-extern {
-    type Location = usize;
-    type Error = ();
-    
-    enum Tok<'input> {
-        "num" => Tok::NumSymbol(<&'input str>),
-        "FactorOp" => Tok::FactorOp(<Opcode>),
-        "ExprOp" => Tok::ExprOp(<Opcode>),
-        "(" => Tok::ParenOpen,
-        ")" => Tok::ParenClose,
-    }
-}
-
-
+Num: i32 = {
+    r"[0-9]+" => i32::from_str(<>).unwrap()*scale,
+};
