+++ doc/src/tutorial/index.md
@@ -9,8 +9,8 @@ first.
 - [Handling full expressions](004_full_expressions.md)
 - [Building ASTs](005_building_asts.md)
 - [Macros](006_macros.md)
-- [Error recovery](007_error_recovery.md)
-- [Passing state parameter](008_state_parameter.md)
+- [Error recovery](008_error_recovery.md)
+- [Passing state parameter](009_state_parameter.md)
 
 This tutorial is still incomplete. Here are some topics that I aim to
 cover when I get time to write about them:
